Add files via upload
Hello guys, this is a simple AFK Wood farm Script for Grand RR.

Requirements:
- You have to be Level 3 as Wood farmer, because of the Task that changes with the level
- The minimap have to be in the left lower corner as usual and the points have to be yellow as usual

Install Python library:
pip install ImageGrab
pip install opencv-python
pip install pynput

And of course you need to have python, i use 3.11.9 and it works so maybe go with it ( https://www.python.org/downloads/ )

You have to get the job and it would be very good if you get in the area where you have to work, then, start the programm, tab into the game and he should be running of his own. Just leave him like that, he will not be as fast as yourself but it works and you can let it run for multiple ours.

I will try my best to update it and improve it over time maybe with electrician job or something else.

Have fun.